Sort Google Drive items using last modification date
There is also a “last viewed” and “last modified by me” date, but I think “last modified” makes the most sense to use for this.

Note though that Alfred will “rank” the items based on filter string, so this can break the ordering, but I _think_ that for items that the user has never selected, Alfred uses the original order.

Closes #19.