Add “check for updates” to the workflow
This will run whenever the user selects an item from the script filter, and if there is an update available, it will show a notification.

Note that this is the only action that will update the cached ‘github-releases.json’ used to learn if we have a newer release.

We could have our main script do this in the background, but then we wouldn’t be able to show a notification (when a new release is found).