Handle “already in use” exception when setting up http server
This happens if the user previously requested items from Google Drive without having an access token and without authenticating.

This will leave our script waiting for the access token and future scripts will fail because they cannot setup a server.

We still need to abort, because we cannot setup a server, and thus we do not know when the user authenticates, but we can redirect them (again) to the browser, where they are asked to authenticate.