Introduce custom_query workflow variable
This allows extending the default query used to get files from Google Drive.

For example to exclude folders from the request, set it like this:

    mimeType!='application/vnd.google-apps.folder'

It should also be possible to limit search to a specific folder by requiring that “parents in «folder id»” but this might require updating to version 3 of the Google Drive API.

Please refer to Google’s documentation about the possibilities. The workflow currently uses version 2 of the API.

Closes #38